[FIX] sale_subscription: display subscription_discount notes on portal
Since https://github.com/odoo/enterprise/pull/73438
sale.order.line with display_type=='subscription_discount' are used to
easily update the line name when the upsell period is updated.

The portal template was not updated. As a result, the line was not
displayed on the portal view of subscription.

closes odoo/enterprise#82751

Taskid: 4690494
Related: odoo/odoo#204498
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>