[IMP] documents: convert emails to documents
For instance, when receiving invoices from digital services, you sometimes
just receive your invoice/receipt as an email without any attachment.

Currently you'd have to open the email, print/save the email as PDF and
either upload it manually to the Documents app, or attach it to an email
sent to a Documents alias.

This commits converts emails without attachments sent to Documents into
documents.

task-4539553

closes odoo/enterprise#82057

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>