[FIX] l10n_be_hr_contract_salary: remove dependent people
When you uncheck the 'Other dependent' box, it keeps the value in the
dependent fields. This leads to have values set on the employees that
are not visible.

To avoid this issue, we empty the 'dependent' fields when we uncheck the
'Other Dependent' checkbox. This is the same behaviour of what is done
in the onchange on backend view.

closes odoo/enterprise#82820

X-original-commit: ead61f4d81a384f45f935524bcdd1dcf3f9194a6
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>