[FIX] website_helpdesk: show menus with helpdesk URLs
**Issue**
Menu items starting with helpdesk in their URL were not dispalyed to the
public user in some cases.

**Steps to reproduce**
- Have `website_helpdesk` installed.
- Activate the website form for a helpdesk team.
- Create a menu item for the website, using a URL starting with
helpdesk, e.g. `/helpdesk-123-test`
- This menu is always invisible to the public user.

**Cause**
This issue was fixed in previous versions, but it appears the
duplicated `_compute_visible` code present in `website.py` was
not deleted in forward ports of the fix.
https://github.com/odoo/enterprise/blob/b320390d23d3f009beb809e0334c8fa61c44d824/website_helpdesk/models/website.py#L24-L26
As a result, the menu item starting with `/helpdesk` is matched
as a helpdesk menu, but will logically not be part of the helpdesk
published menus:
https://github.com/odoo/enterprise/blob/b320390d23d3f009beb809e0334c8fa61c44d824/website_helpdesk/models/website.py#L30-L32

Note: also prevent creating a `/helpdesk` menu item, which is already
used by the helpdesk feature.

opw-4560291

closes odoo/enterprise#84084

X-original-commit: 0d974a9bec88f2c17f1a3d5c09bf93ec0af2b609
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>