[IMP] l10n_mx_edi: Improve 'objetoImp' / tax objected on CFDI
Since the launch of CFDI 4.0 we have had the option to check a box and not have the taxes displayed on the CFDI.
That allows to have control of objetoImp = 03.
However, there are 8 values to cover and only 3 are managed. Also, the user has no control of it.

task_id: 4365578

closes odoo/enterprise#82903

Related: odoo/upgrade#7708
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>