[IMP] timesheet_grid: avoid user spamming timer button
Before this commit, the user was able to start/stop and start again very
quickly which crashes the view because the timer has not the time to
correctly be loaded in the view.

This commit adds a delay to avoid having that issue.

task-4284216

Part-of: odoo/enterprise#81686
Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>