[FIX] documents: keep documents actions visible
1/

1. Select a record in kanban view in a folder with a few
  records (< page size)
2. Switch to list view
3. The documents should still be selected, some actions are
  visible, but not all

This occurs because the old view is unmounted only after the
new view is already mounted, so the service function is set to
null.
We don't need to clear it when unmounting because no other
context (not a documents view) is expected to call this method
on the documents service.

2/

1. Select a pdf record in kanban view
2. Open Split Tools
3. Actions are not shown

Task-4702653

closes odoo/enterprise#83501

X-original-commit: 9830aed2a6b056f737244de421e84fd52c2b4491
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>