[IMP] mrp_workorder: shop floor stop workorders on employee logout
In the old employees panel of the shop floor, users could click on the
'x' button next to an employee to remove them from the list, and to stop
the timer for this employee on all workorders.
With the new Edit Operators popup, this was no longer the case.

To resolve this, this commit now ensures that all the users that were
logged out using the Edit Operators panel have all of their current
timers on workorders stopped. All cards are also refreshed upon closing
this dialog to ensure that any stopped timers do not appear to still be
running in the shop floor itself.

task-4484941

Part-of: odoo/enterprise#80469
Related: odoo/odoo#200345
Related: odoo/upgrade#7331
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>