[IMP] account_accountant, l10n_mx_edi: Refactor _action_reload_liquidity_line() as a context manager
Due to the cache invalidation that occurs in order to force the recomputation of computed fields, we need to track the statement line in the context.
Previously, this was managed manually in a stable-version fix:
https://github.com/odoo/enterprise/commit/46f838403c8d60760a5b37cb5fe8d052d2307999
This commit improves the approach by refactoring _action_reload_liquidity_line() into a context manager.

no-task

closes odoo/enterprise#82819

Signed-off-by: Laurent Smet (las) <las@odoo.com>