[FIX] account_sepa_direct_debit: print SDD mandate payment info
Steps to reproduce:
- Install Contacts, Accounting and l10n_be
- Switch to a Belgian company (e.g. BE Company CoA)
- In Accounting settings, activate "SEPA Direct Debit (SDD)"
- Create a contact:
  * Country: Belgium
  * Bank Account: [a Belgian IBAN]
- Go to "Accounting / Customers / Direct Debit Mandates"
- Create a mandate for the created contact
- Validate the mandate
- Create an invoice for the created contact as customer
- Confirm the invoice
- Pay the invoice with "SEPA Direct Debit"
- Print the invoice

Issue:
The information explaining that the invoice has been paid using direct debit
and that the invoice should not be paid manually is not printed on the invoice.

Cause:
This information is printed if "sdd_mandate_id" field of the invoice is set.
"sdd_mandate_id" is a related field to "origin_payment_id.sdd_mandate_id"
field.
However, since the refactoring on "account.payment", "origin_payment_id" is
not set on an invoice anymore.
It is just set on the journal entry linked to a payment.
Therefore, we cannot rely on that field anymore to print the SDD mandate
payment information.

Solution:
The "matched_payment_ids" field can be used instead.
It's a Many2many field referring all payments linked to the invoice.
If one of these payments has "sdd_mandate_id" set, the SDD mandate payment
information is printed on the invoice.

opw-4672191

closes odoo/enterprise#82929

X-original-commit: 34ea804075139de1e86949f13340a37bc2a091c0
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>