[FIX] account_bank_statement_import_camt: fix import of CAMT file
Steps to reproduce:
- Accounting dashboard
- Click on the three dots of the "Bank" card
- Click "import file"
- Try importing a CAMT file with the namespace "camt"
- Traceback

- Try importing a CAMT file with a custom Code in <Cd> (example: "IC1")
- Traceback

Cause:
Namespaces can be used in CAMT files but the way Odoo handled them was incorrect.
We were creating a dictionary using `root.nsmap`. It only worked when the file didn't have a namespace key because if the key was different from 'ns' the `findall` didn't work.

Second issue: Some banks use their own codes in the CAMT files. The specification has numerous codes which Odoo added in a dictionary, and it reads from this dictionary to get a description. When there is a custom code, Odoo tries to read a key which doesn't exist explaining the traceback.

Solution:
- Change the way namespace is computed by querying the namespace of `root` instead of using `nsmap`.

- If the code is not in the dictionary, we take the custom code.

opw-4553152

closes odoo/enterprise#84329

X-original-commit: e721bc2ef0d312612760582c0178427beb075262
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>