[FIX] approvals_purchase: correctly update product seller
-- Context --
When no seller is defined on the product that is included in the approval
request, you cannot generate an RFQ. Adding a seller in the product should
allow creating the RFQ without re-setting the request to draft.

-- Behavior before the fix --
You had to reset the request to draft before being able to generate the
RFQ when a seller has been added to the product.

-- Behavior after the fix --
You can. There was a missing field in the depends of the compute method
for the seller ids.

task-4680780

closes odoo/enterprise#83991

X-original-commit: f267532689912eaf43e8b3d1dc5ba49f62b88026
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>