[IMP] payment_sepa_direct_debit: create token on mandate confirmation
Prior to this commit:

When an sdd.mandate is confirmed in the backend and a SEPA provider is
active, it does not create a payment token. It means, it's not
possible to assign that token to a subscription.

Post this commit:

When an sdd.mandate is confirmed, if there is an active SEPA provider
for the right company and the right journal, a token is created.

task-4100555

closes odoo/enterprise#73543

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>