[IMP] planning: multi create in calendar view of planning
Before this commit, it was not possible to allow the creation of a bunch
of planning slot in the calendar view. However, thanks to a new feature
implemented by payroll team for the calendar view of work entries, it is
now possible to create a bunch of records in the calendar view.

This commit adds the multi-create feature in calendar view in the
calendar of planning slot model. The user just has to select a planning
template and the resource to assign on the new shifts and then select
the days to create shifts.

task-4485911

closes odoo/enterprise#84041

Related: odoo/upgrade#7918
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>