[FIX] account_accountant: reco widget with branch error
Steps:

- Have a main company C with a branch B
- Select B in company selector, but have C checkbox checked too
- Create a user U with `company_id == B`
- Create and confirm an invoice I for user U, company_id should be B,
  amount X
- Go back to accounting dashboard, and open reconciliation widget
- Create a new transaction, amout X
- Select it, select line from I and click on Validate
-> UserError: "Incompatible companies on records"

This reason is, even if we have Branch company selected as
`env.company`, as we have Main company checked too, the statement
line is created with `company_id == main company`

With this commit, we unset the partner from the statement line's move
if it is from an accessible branch but not the main company.
We also test different usecases to secure the flow.

opw-4626664

closes odoo/enterprise#85431

X-original-commit: 311dcc223e3c9b592d0203f6929100fa47a6141c
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>