[IMP] documents(,_sign): add inbox folder
Creation of Inbox folder, taking the inbox alias name
unless present on Internal (or other) in existing db.
This is why the alias is a separate record instead of
configured through alias_name on the folder.
See upgrade PR.

Embed `Send to Finance` and `Sign` actions on it (if
module is present).

Task-4690653

Part-of: odoo/enterprise#82891
Related: odoo/upgrade#7846
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>