[FIX] l10n_it_riba: use payment date instead of invoice due date
Partial revert of the odoo/enterprise#78827 fix.

The payment due date is in facts the payment date itself.
This way we are able to handle different lines each with a
different due date.

The user will have to put it correctly manually in the Register
Payment wizard even if the Payment Terms are specified on the invoice.

opw-4703520

closes odoo/enterprise#83014

X-original-commit: 450d7b3d867db6a32321ed8034e9e0c5f750ec16
Signed-off-by: William André (wan) <wan@odoo.com>