[FIX] marketing_automation_website_sale: anniversary campaign error
Try creating the anniversary campaign, it will throw an error.
The reason was due to a recent change in `ir.actions.server` [1].
This commit [2], removes the default value of `state`, but we're not passing
any value for `state` while creating the campaign, thus it throws
an error, as `state` is required for server action.

[1]: https://github.com/odoo/odoo/pull/202347
[2]: https://github.com/odoo/odoo/commit/e84210e5b2de8791ced9ee715197a6b7db26eecb

Task-4707098

closes odoo/enterprise#82953

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>