[FIX] l10n_mx_edi_stock: Set l10n_mx_edi_stock fields
In databases migrated to Odoo 18, that has the module l10n_mx_edi_stock installed the field l10n_mx_edi_hazardous_material remains empty. This issue occurs because the CSV file (product.unspsc.code.csv) assigns the value but the records are no update

As a result, customers cannot generate the Carta Porte in the l10n_mx module, encountering the following error:

Error 301 (CP155):
The attribute "Mercancia:MaterialPeligroso" is missing or invalid when the ClaveProdServ is declared as hazardous.

closes odoo/enterprise#83229

X-original-commit: 3dc07b926d770ef6b3db8d22792de7144f215f6d
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>