[FIX] l10n_mx_hr_payroll_localisation: hourly wage
Steps to reproduce:
- Install payroll, payroll Mexican localisation
- Switch to the Mexican company
- Choose an employee from the Mexican company
- Change the contract to be hourly waged instead of monthly
- Set a value for the wage/hour
- Create some work entries for that employee
- Generate a payslip for the employee in the duration you are testing
- Edit the work entries you have to have some unpaid leaves
- Generate another payslip for the employee

Current Behavior:
Payslip is generated without taking into consideration the number of
hours actually worked and the unpaid leaves taken by the employee. It
also generates the payslip with the value of the wage/month regardless
of the value of the wage/hour you added.

Issue:
In df8f53b0cad4738537f18890e628c286dedaa141, Mexican localisation was
implemented but with computing the amount of the `payslip_worked_days`
using the `contract.wage` which is the fixed amount of wage/month.

https://github.com/odoo/enterprise/blob/25caeeab01abc0b47f6bd6b02658c3bbbe0aba96/l10n_mx_hr_payroll_localisation/models/hr_payslip_worked_days.py#L19-L20

opw-4604675

closes odoo/enterprise#84715

X-original-commit: 92f48d4df3622aa4f420e599b1cf4cf1d96574d0
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Hussein Yasser Hussein Ebrahim (hyhe) <hyhe@odoo.com>