[FIX] l10n_ar: make use of invoice_currency_rate.
PR https://github.com/odoo/enterprise/pull/82924 makes use of the field
l10n_ar_currency_rate. But that field is not present in Odoo versions above 17.0

This field was removed in favor of invoice_currecy_rate from account_move, that
stores the currency rate used for a document in Odoo 17 and above.

This fix adapts the l10n_ar module to use that field instead.

opw-4708505

closes odoo/enterprise#83053

X-original-commit: c0ed3a9f0c43ab29b7de94242ee704ebf19da3ad
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>