[IMP] appointment,*: transform questions into a m2m field
*: appointment_account_payment,pos_restaurant_appointment,
website_appointment_sale_project

PURPOSE
=======

In order to make the question handling easier when using the
appointment module regularly, especially with similar appointment
types, we want to move from appointment-specific questions (o2m)
to a library of questions that can be used in all appointments.

AFTER THIS COMMIT
=================

The question_ids field is now a m2m, so that questions can be
customized and translated a single time and reused across appointments.

A new field is_reusable is added on the appointment.question model
in order to allow the user to configure whether the question will
be available for selection in the list of existing questions when
adding one in the question_ids field of an appointment_type. The
field is True by default.

OTHER CHANGES
=============

The form view of the appointment.question is slightly enhanced using
a new widget based on the mail badge widget, but improved with
a parameter for fa icons before each of the selection field values.
A js test is added.

We also add menu 'Answers' in reporting and 'Questions' in the
configuration. We also now order questions by id in addition to
sequence, to improve determinism.

Move answer views to separate file and make the naming match the
guidelines. Also use a single main action for answer inputs that
is modified in the python action code.

Make fields readonly on the answer input form view except text and
selected values. On the calendar event, make question readonly.

Filters are added in the search view of appointment.question for
is_reusable: 'Reusable' and 'Not Reusable'.

Tests are updated according to the changes.

UPGRADE
=======

- We create a new column for is_reusable. Note that we do not set
a default True there because we want all existing questions to be
'not reusable' after migration, while new questions, once the
feature is present, will be reusable by default.
- We rename the reporting menu id.
- We remove questions without a type as it is now made mandatory.

See UGR PR: https://github.com/odoo/upgrade/pull/7691

Task-4655040

Part-of: odoo/enterprise#82973
Related: odoo/upgrade#7691
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>