[FIX] l10n_cz_reports_2025: Convert invoice amounts to company currency in VIES summary report
**Issue**
When an invoice is issued in a currency different from
the company currency, the VIES summary report incorrectly
displays the total value as if it were already converted.
This results in incorrect totals in the report.

**Steps to Reproduce:**
1. Install the Accounting app and the l10n_cz_reports_2025 module.
2. Go to Accounting > Customers > Invoices.
3. Create a new invoice using a currency other than CZK.
4. Add a product with Code Supply set to Gold and Transaction Code
   set to 0 Goods. Confirm the invoice.
5. Navigate to Accounting > Reporting > VIES Summary Report.
6. Observe that the total value is not correctly converted
   to the company currency.

**Expected Behavior:**
The total amounts in the VIES summary report should be
correctly converted to the company currency if the invoice
was created in a foreign currency.

**Actual Behavior:**
Amounts are displayed in the invoice currency without
conversion, leading to incorrect totals when the invoice
currency differs from the company currency.

**Root Cause**

The report fetches raw invoice data without handling
currency conversion. Specifically, it retrieves the unconverted
total value, even when the invoice currency differs from the company’s.

**Fix**
To solve the issue the balance field is used, which
correctly reflects the value in the company currency.

opw-4688638

closes odoo/enterprise#98969

X-original-commit: 17f633b30c454565d7177d5169093ec7bfb59652
Signed-off-by: Carmine Santomauro (csan) <csan@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>