[IMP] hr_payroll: allow deleting unused structures
Currently, deleting unused structures is not straightforward as it requires
to delete all attached salary rules first. It would be easier to automatically
delete all salary rules of the structure upon deletion. This operation will fail
in case the salary structure is used in at least one payslip, which is what we would expect.

This commit allows salary structures to be deleted as long as they are
not used anywhere.

task-4702289

closes odoo/enterprise#83020

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>