[FIX] sale_amazon: don't manage unused references when sync order
Following commit 6b5acdde9998fb1ee0d4d4aee0ab2375e50d7703 we ensured
that we stopped the stock synchronization when there was a discrepancy
between an offer fulfilment channel and an order fulfilment channel.

This bugfix didn't consider the cases of offers that were synced once
then never updated with the changes of the new API, thus not being the
string of a dictionary, but instead, being the string of an int. Thus
raising an attribute error, and not a JSON error nor a TypeError.

These offers, though, don't need that safety net, and can thus just be
ignored. If their stock needs to be sync someday, the stock cron will
take care of these.

opw-4684657
opw-4684605

closes odoo/enterprise#83117

X-original-commit: a3d29f62567afd77a862fb1da21bab4e82d48cb1
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>