[FIX] knowledge: use correct model for permission tests
This commit changes a test that checks that we cannot remove a specific member
from an article. The issue was that the parameter provided didn't respect the
function's signature by giving a `res.users` instead of a
`knowledge.article.member`. By chance, it always worked as those model share the
field partner_id that represent the same data. But it is updated so that it
better represents the correct use of the method.

task-4712569

closes odoo/enterprise#83243

X-original-commit: 6fbb5bb732750254e1bdf2a0417830bc2e0df1b3
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>