[IMP] helpdesk_account: allow multiple refunds per invoice in helpdesk flow
Previously, users couldn't select an invoice for another refund if all of its
credit notes were already paid or in payment — even if the invoice was only
partially refunded.

For example:
If a user created a partial refund for the first time and marked it as paid, the
original invoice could no longer be selected for any further refunds.

Now, invoices remain available for refunds as long as the total refunded amount
is less than the original invoice's amount_total.

Task ID: 4582294

closes odoo/enterprise#83125

Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>