[FIX] hr_payroll: Aggregate inputs amounts if rule code not matching
Purpose
=======

Since: https://github.com/odoo/enterprise/commit/65b4e343ce9b16ef77f6c967314c48d0b10972bc
The amounts are aggregated outside of the rule enumeration in the case there are
several inputs for which the code is matching the salary rule code.

But in the case the rule code is not the same as the other input code, only the last
amount is retrieved, not the total amount.

Taskid: 4711431
X-original-commit: 778071c8742ab51b96497a752ebb2b682fbc7385
Part-of: odoo/enterprise#83121
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>