[FIX] account_online_synchronization: fetching from branch companies
If you have a company A that has a journal set up with bank synchronization,
and a company B that is a branch of company A, you have access to the journal
and the bank synchronization (account.online.link) of company A in company B.

This causes an issue when fetching the transactions from the context of
company B, as that will create the fetched transactions with a `company_id`
set to B, which will cascade on the journal entries and items linked to these
transactions.

To ensure the correct company is set on the transactions, we force the company
in the context when creating them.

As a result:
- When only company A is selected, transactions are correctly created in
company A.
- When both company are selected, no matter if the transactions are fetched from
A or B, the transactions are correctly created in company A.
- When only company B is selected, the user is faced with an access error, as
Odoo is trying to open a view with the fecthed transaction, which belong to
company A.

To prevent this access error, it was decided that the "Fetch Transaction"
buttons should not appear when the company owning the bank synchronization is
not selected.

A new non-stored computed field `show_sync_actions` was then introduced on the
account.online.link to check if we should allow fetching given the currenctly
selected companies.

This field is then used to decide whether if we show the transaction fetching
buttons on the dashboard and on the online link form view.

Another issue with branches is the 'connect bank' button.

With the same configuration as above, let's have companies A and B enabled,
with B as the selected company.

If you connect a bank to a journal belonging to company A by clicking on the
'connect bank' button on the dashboard, it will successfully connect, but the
account.online.link will belong to company B.
Then, if you try to access the accounting dashboard with only company A
enabled, you will be faced with an access error, as the journal from company A
tries to diplay informations computed with the above account.online.link of
company B.

To prevent this, we only make the 'connect bank' button appear if company A
(i.e. the company owning the journal) is selected as the active company.

This did not require a new field, and was done by adding an entry in the
dashboard data.

A test was added to check the value of the new field/dashboard data entry with
different combination of selected/enabled companies in a similar configuration
as we used above with A and B.

opw-4515862

closes odoo/enterprise#84830

X-original-commit: 9b133b2cd03a726d330c87a1227903fc925b0ed5
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florent Huylenbroeck (flhu) <flhu@odoo.com>