[IMP] base: upstream sync partner address
It is now possible to update address of a contact. When doing so
we actually update the parent address. Rationale is that: if you
edit the contact of a company, you actually edit the company, at
least for what concerns the address. Views are updated accordingly.

See community PR for more details.

Task-4708358
Part of Task-4377720

Part-of: odoo/enterprise#83446
Related: odoo/odoo#206112
Related: odoo/upgrade#7564
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>