[FIX] documents_account: use correct accounting wording
TLDR: reword "Create Vendor Refund" => "Create Vendor Credit Note"

During the refactoring in
https://github.com/odoo/enterprise/commit/a32825ee00f2b330d99113f4d8c1488903fe744e

The enum value of "account.move.in_refund" was used to name the button
"Create Vendor Refund" instead of its string "Vendor Credit Note" when
converting the action into the new documents actions scheme. This is
confusing for customers (since it is not always a refund) and is
inconsistent with the rest of the user facing wording (i.e. the form
view that opens when the button is clicked on).

Reported/noticed during:

opw-4488038

closes odoo/enterprise#83130

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>