[FIX] account_bank_statement_import_camt: handle more multi-currency rate
The current implementation to handle the rate for multi-currency entries
in CAMT creates problems when fetching the source rate from the file:
the source rate can be taken from the SrcCcy or the TrgtCcy node, which
should be compared with the source currency and target currency
respectively. It is not done that way and is too restrictive (when the
SrcCcy is found, it is always compared to the target currency, which
leads to an incorrect transaction amount).

Moreover, when the currency conversion introduces a exchange loss/gain,
it is put back on the largest transaction of the entry.

Backport of https://github.com/odoo/enterprise/pull/76828 and
https://github.com/odoo/enterprise/pull/42763

opw-4672283

closes odoo/enterprise#86848

X-original-commit: 830ca6c84c807dba5a405f5d48e515b867cd24f3
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>