[IMP] project_timesheet_forecast: only take record in past or in today
Before this commit, the Planning / Timesheets analysis take into account
the record in the future. It could be okay for the planning shifts but
for the timesheets, it does not make any sense since the timesheets are
only recorded when the work is done.

This commit filters the records retrieve inside that report to make
sure we only take the records in the past or today.

task-4297336

closes odoo/enterprise#79043

Related: odoo/odoo#205503
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>