[FIX] documents{_account,}: allow previewing shared documents
Currently, public users cannot preview shared documents; instead, the
documents are always downloaded.

Steps to reproduce
------------------
1. Install `documents_account`.
2. Share a previewable document (e.g., an image or a PDF) with a link
   that allows public viewing.
3. Open the share link in an incognito window.
4. Click "Preview file."

The file is downloaded instead of being previewed.

Cause
-----
In an override of the route handling document content retrieval, the
argument that determines whether the document should be displayed or
downloaded is lost in a `super()` call.

opw-4490958

closes odoo/enterprise#83254

X-original-commit: 453fdf9ca4f1e988f5f938fe01b7cce5e7c88d57
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>