[IMP] l10n_{co,mx}_edi: Grouping key for base lines with no tax
At the moment, the tax aggregators always give a base line with no tax
the `None` grouping key.

Sometimes, we want base lines with no tax to be grouped as if they had
a particular tax, such as an exempt one.

This commit extends the aggregator behaviour so that for base lines with
no tax, the grouping function is called with an empty tax_data dict, and
the returned grouping key is used to group the base line.

closes odoo/enterprise#88038

X-original-commit: 1935a40b15f5f31b13621faf55be3c3821f07434
Related: odoo/odoo#214982
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>