[FIX] sale_subscription: make payment cron resilient to interruptions
The is_invoice_cron flag was set at the start of the cron and only reset
at the end. In addition, key processing steps (_process_invoices_to_send
and _post_invoice_hook) were done after all subscriptions were processed.

If the cron was interrupted, unprocessed subscriptions could be left
with is_invoice_cron set to True indefinitely, blocking post-processing.
This also caused the post-processing cron to accumulate a backlog of payments to handle.

To fix this, subscription processing is completly done in the loop.
 the is_invoice_cron flag is set only when a subscription (or group) is being processed and
cleared immediately afterward. We cannot remove the commit inside the
loop due to payment processing but at least if the cron is interrupted
only one subscription end up in an intermediate state

This ensures that only a single group may be left flagged if interrupted,
and the flag will be reset when the cron resumes.

closes odoo/enterprise#83868

X-original-commit: daac4da302e63f87fe3f244099e26b47d08d7cfb
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Thibault Francois <tfr@odoo.com>