[FIX] l10n_mx_reports: filter account lines correctly
Currently, an error may occur when we try to export the trial balance
with the SAT (XML) format. This is because a None id is used
to browse the account, due to a more specific account_group.

Steps to reproduce:
- Install l10n_mx_reports.
- Create an account_group with a more specific account, such as 401.01.01
    - The 4, 401, and 401.01 groups are created with demo data.
- Go to trial balance and try exporting with "SAT (XML)".

An error will occur (`AttributeError: 'bool' object has no attribute
'startswith'`) because the account is None, so account have no type.

opw-4716047

closes odoo/enterprise#83594

X-original-commit: ff726c5eb879517eaf9754d8fa6269b52924cf7a
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>