[FIX] hr_payroll: prevent error when adding a line to rule parameters
The error occurs when attempting to add a date in line in the Salary Rule
Parameters

Steps to Reproduce:
- Install the 'hr_payroll' module
- Payroll > Configuration > Rule Parameters
- Create New > History > Add a date

Traceback:
AttributeError
'bool' object has no attribute 'replace'

At [1], this error occurs because 'value_id.parameter_value' is False. This
happens because first we're adding a date, so 'value_id.parameter_value' is
initially empty.

[1]- https://github.com/odoo/enterprise/blob/b52a93da7d680b60646cb08a9126cf6c12f5307d/hr_payroll/models/hr_rule_parameter.py#L101-L102

sentry-6531613276

closes odoo/enterprise#84092

X-original-commit: 09ff00c528bfa6df66d0e15c834c67ee54a472a4
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>