[FIX] web_gantt: make sure orm service is defined in gantt renderer
Before this commit, when the user reschedules his task with dependencies
by clicking on any arrow displayed in the gantt connector. A
notification is displayed to notify the user the action is done with
success, that notification also contains an `Undo` button to let the
user to undo the rescheduling made. However, if the user clicks on the
undo button, a traceback occurs. The reason is because `orm` service is
not defined in the gantt renderer.

This commit makes sure the orm service is defined in the gantt renderer
to be able to call the `action_rollback_scheduling` method.

task-4687572

X-original-commit: 60a6ef104956ae6e8740c59680dee84e5913b900
Part-of: odoo/enterprise#83297
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>