[FIX] web_gantt: don't display undo button if old vals are not given
Before this commit, after the `web_gantt_schedule` call was done, a
notification is displayed according to the result received by the
`web_gantt_schedule` call done. The problem is even if the
`old_vals_per_pill_id` is not given when the gantt scheduling has been
made with success, the `Undo` button is displayed. The problem is
without `old_vals_per_pill_id`, it is impossible to undo the action
done, because the initial data before the `web_gantt_schedule` is not
known without that data, and so, if the user clicks on `Undo` button a
traceback will occur since we cannot retrieve the initial data of the
records processed in the `web_gantt_schedule` method.

This commit makes sure the `Undo` button is displayed only if we have
the information to `undo` the `web_gantt_schedule` method. To know that,
the `old_vals_per_pill_id` should be given in the result of
`web_gantt_schedule` to be able to revert the changes made by
`web_gantt_schedule`.

task-4687572

X-original-commit: bab936e61bf197deadb99eeda28708561167b4bd
Part-of: odoo/enterprise#83297
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>