[FIX] web_gantt: correctly display warning notification
Before this commit, a warning notification was not properly displayed
because the method does not return what we expect. Instead of returning
the message and notification type, it returns a client action to call
`display_notification`. Moreover, the warning notification does not
display the right icon, it display `fa-check` instead of `fa-warning`.

This commit changes the object returned for the warning to have the
expected values returned and changes the icon when the notification is a
warning one.

task-4687572

closes odoo/enterprise#83297

X-original-commit: 94322474109d60534fdeaaf8ea9ee6c98ab377f9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>