[FIX] documents{_project}: upload file on correct request document
Step to reproduce :
  - Create a folder with a sub folder and a request or two requests.
  - Share the parent folder with edit permission to anyone.
  - Go to the public folder view.
  - Upload a file for the last request document.
  - The file will be uploaded to the first editable folder/request found.

This was due to a hidden form tag to upload file to sub folders which was using
the same class as the one used by the request documents (o_request_upload).
Using class instead of id also means that it wasn't possible to share more than
one request document.

Task-4718126

closes odoo/enterprise#85103

X-original-commit: 495db49f7f06148b91a4ff177642c41197be8df2
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>