[FIX] whatsapp: prevent error when receiving file in log message
When Debug requests is enable on the whatsapp business account and
user replies with the file, a traceback will appear.

Steps to reproduce the error:
- Install "whatsapp" and "contacts"
- Create a whatsapp business account > Enable Debug requests
- Create a contact > Add a phone number > send whatsapp message with
  any whatsapp template
- Now reply with the file(ex. pdf or image) in message from
  the contact's whatsapp.
- Traceback in terminal

Traceback:
"ValueError: A string literal cannot contain NUL (0x00) characters."

This is because logging `response.text` makes no sense for non-text
values. Instead for octet-stream responses we want to log the size
of the contents.

Some accomodations are made to allow testing http response handling
inside the "request" part of the API that's normally disabled
during testing.

Logging is moved out of the request try block, as there is no
handling in cases where it would fail and we should not expect
it to fail anyway.

sentry-6314521165

closes odoo/enterprise#91138

X-original-commit: 36e5b8125781642e0b49b6cf1aaaa004ba43236a
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>