[PERF] iap_extract: remove savepoint when uploading to the OCR server
As this savepoint was initiated in a loop, it could lead to many
subtransactions (one savepoint per document to upload) being active at
the same time, causing performance issues (more info at [1]).

The main goal of that savepoint was to be able to recover from a SQL
serialization failure that could occur when the document to upload was
modified concurrently.

This was mostly useful in version 16.0 where the upload of the documents
to the OCR server was performed asynchronously through a cron. This
greatly increased the likelihood of serialization errors.

In version 17.0, we went back to synchronous uploads (commit 9df91d8),
these should be much less frequent.
Considering this, it seems reasonable to remove the savepoint, although
it implies that if such error occurs, the SQL transaction will be
rollbacked and some OCR credits might be lost in the process.

[1]: https://www.postgresql.org/docs/current/subxacts.html, in
particular the last sentence:
> The more subtransactions each transaction keeps open (not rolled back
> or released), the greater the transaction management overhead. Up to
> 64 open subxids are cached in shared memory for each backend; after
> that point, the storage I/O overhead increases significantly due to
> additional lookups of subxid entries in pg_subtrans.

closes odoo/enterprise#83478

X-original-commit: 990c897f4ca7e8506707427b43b54bd8b4734433
Signed-off-by: Olivier Dony (odo) <odo@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>